chore: add Dockerfile and docker-compose

+++ backend/voip/main.py
@@ -1,3 +1,5 @@
+import io
+import os
 import subprocess
 import time
 
@@ -7,7 +9,8 @@ from voip.models import VoIPConfig
 
 def perform_call(config: VoIPConfig, trigger: Trigger, receiver_phone: str):
     call_params = [
-        ".\\voip\\Caller.exe",
+        "dotnet",
+        "Caller.dll",
         str(config.username),
         str(config.password),
         str(config.sip_server_address),
@@ -15,7 +18,17 @@ def perform_call(config: VoIPConfig, trigger: Trigger, receiver_phone: str):
         str(trigger.message_file.path),
         str(receiver_phone),
     ]
-    process = subprocess.Popen(call_params)
+    process = subprocess.Popen(call_params, cwd="voip", stdout=subprocess.PIPE)
     print(call_params)
+
     time.sleep(40)
+
+    result = ""
+    for line in io.TextIOWrapper(process.stdout, encoding="utf-8"):
+        result += line
+        print(line)
+
+    if "InCall" not in result and "Completed" not in result:
+        raise Exception('Unsuccessful call')
+
     process.kill()
